# internal/docs/swagger.yaml
basePath: /
definitions:
...
  dto.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  dto.TaskListResponse:
    properties:
      tasks:
        items:
          $ref: '#/definitions/dto.TaskResponse'
        type: array
    type: object
  dto.TaskOptionRequest:
    properties:
      is_correct:
        type: boolean
      text:
        type: string
    type: object
  dto.TaskOptionResponse:
    properties:
      id:
        type: string
      is_correct:
        type: boolean
      text:
        type: string
    type: object
  dto.TaskRequest:
    properties:
      correct_text_answer:
        type: string
      course_id:
        type: string
      module_id:
        type: string
      options:
        items:
          $ref: '#/definitions/dto.TaskOptionRequest'
        type: array
      title:
        type: string
      type:
        type: string
    type: object
  dto.TaskResponse:
    properties:
      correct_text_answer:
        type: string
      course_id:
        type: string
      id:
        type: string
      module_id:
        type: string
      options:
        items:
          $ref: '#/definitions/dto.TaskOptionResponse'
        type: array
      title:
        type: string
      type:
        type: string
    type: object
...
info:
  contact: {}
  description: API для пользователей Course Service
  title: Course Service API
  version: "1.0"
paths:
...
  /tasks:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные задания
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать задание
      tags:
      - Tasks
  /tasks/{id}:
    delete:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить задание
      tags:
      - Tasks
    get:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить задание по ID
      tags:
      - Tasks
    put:
      consumes:
      - application/json
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Новые данные задания
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.TaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Обновить задание
      tags:
      - Tasks
  /tasks/courses/{course_id}:
    get:
      parameters:
      - description: Course ID
        in: path
        name: course_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить задания курса
      tags:
      - Tasks
  /tasks/modules/{module_id}:
    get:
      parameters:
      - description: Module ID
        in: path
        name: module_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TaskListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить задания модуля
      tags:
      - Tasks
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"